chore: refactor tests and use Jasmine.
logion-network/logion-internal#837
