+++ packages/node-api/package.json
@@ -19,7 +19,7 @@
     "build:cjs": "tsc -p tsconfig-cjs.json",
     "build:types": "tsc -p tsconfig-types.json",
     "lint": "yarn eslint src/**",
-    "test": "yarn node --experimental-vm-modules $(yarn bin jest)",
+    "test": "NODE_OPTIONS=--loader=ts-node/esm jasmine --config=jasmine-unit.json",
     "integration-test": "docker-compose up -d && NODE_OPTIONS=--loader=ts-node/esm jasmine --config=jasmine-integration.json ; docker-compose down",
     "generate:defs-meta": "yarn generate:defs && yarn generate:meta",
     "generate:defs": "node --loader ts-node/esm ../../node_modules/.bin/polkadot-types-from-defs --endpoint ws://localhost:9944 --input ./src/interfaces/ --package @logion/node-api/dist/interfaces",
@@ -52,16 +52,12 @@
     "@polkadot/typegen": "^10.4.1",
     "@tsconfig/node18": "^1.0.1",
     "@types/jasmine": "^4.0.3",
-    "@types/jest": "^29.4.0",
     "@types/node": "^18.6.1",
     "@typescript-eslint/eslint-plugin": "latest",
     "@typescript-eslint/parser": "latest",
     "eslint": "^8.20.0",
     "jasmine": "^4.3.0",
     "jasmine-spec-reporter": "^7.0.0",
-    "jest": "^29.4.1",
-    "jest-jasmine2": "^28.1.3",
-    "ts-jest": "^29.0.5",
     "ts-node": "^10.9.1",
     "tsconfig-paths": "^4.1.2",
     "typescript": "^4.7.4"
