+++ packages/node-api/src/Connection.ts
@@ -14,8 +14,14 @@ import { LocBatch } from './LocBatch.js';
 import { UUID } from './UUID.js';
 import { LegalOfficerCase, VerifiedIssuerType } from './Types.js';
 
+/**
+ * @deprecated use LogionNodeApiClass
+ */
 export type LogionNodeApi = ApiPromise;
 
+/**
+ * @deprecated use buildApiClass()
+ */
 export async function buildApi(endpoint: string | string[]): Promise<LogionNodeApi> {
     const provider = buildProvider(endpoint);
     const types = Object.values(definitions).reduce((res, { types }): object => ({ ...res, ...types }), {});
