--- packages/node-api/test/Balances.test.ts
@@ -1,41 +1,0 @@
-import { mockPolkadotApi } from "./__mocks__/PolkadotApiMock.js";
-import { PrefixedNumber, ATTO, NONE } from '../src/numbers.js';
-
-mockPolkadotApi();
-const { ApiPromise } = await import('@polkadot/api');
-const {
-    getAccountData,
-    getBalances,
-} = await import('../src/Balances.js');
-
-test("Getting account data", async () => {
-    const api = new ApiPromise();
-    const accountId = "accountId";
-
-    const data = await getAccountData({
-        api,
-        accountId
-    });
-
-    expect(data.available).toBe("42");
-    expect(data.reserved).toBe("0");
-});
-
-test("Getting balances", async () => {
-    const api = new ApiPromise();
-    const accountId = "accountId";
-
-    const data = await getBalances({
-        api,
-        accountId
-    });
-
-    const expected = new PrefixedNumber("42", ATTO);
-    expect(data[0].balance).toEqual(expected);
-    expect(data[0].coin.id).toBe("lgnt");
-    expect(data[0].level).toBe(0.42000000000000004);
-
-    expect(data[1].balance).toEqual(new PrefixedNumber("0", NONE));
-    expect(data[1].coin.id).toBe("dot");
-    expect(data[1].level).toBe(1);
-});
