--- packages/node-api/test/Codec.test.ts
@@ -1,24 +1,0 @@
-import { toHex, fromHex} from '../src/Codec.js';
-
-test("toHex", () => {
-    const data = "test";
-    const hex = toHex(data);
-    expect(hex).toBe("0x74657374");
-});
-
-test("fromHex", () => {
-    const hex = "0x74657374";
-    const data = fromHex(hex);
-    expect(data).toBe("test");
-});
-
-test("fromHex and toHex are compatible", () => {
-    const data = "test";
-    const newData = fromHex(toHex(data));
-    expect(newData).toBe(data);
-});
-
-test("fromHex without prefix fails", () => {
-    const hex = "74657374";
-    expect(() => fromHex(hex)).toThrow();
-});
