--- packages/node-api/test/Vault.test.ts
@@ -1,13 +1,0 @@
-import { mockPolkadotApi } from "./__mocks__/PolkadotApiMock.js";
-import { DEFAULT_LEGAL_OFFICER, ANOTHER_LEGAL_OFFICER } from "./TestData.js";
-
-mockPolkadotApi();
-const { getVaultAddress } = await import("../src/index.js");
-
-describe("Vault", () => {
-
-    it("generates expected Vault address", ()=> {
-        const vaultAddress = getVaultAddress("5H4MvAsobfZ6bBCDyj5dsrWYLrA8HrRzaqa9p61UXtxMhSCY", [ DEFAULT_LEGAL_OFFICER, ANOTHER_LEGAL_OFFICER ]);
-        expect(vaultAddress).toEqual("5HDgYTbcXeXsWxDsdRVKVeBJRgUrq5fPVxRpNaVZYgHkvuzU");
-    })
-})
