--- packages/node-api/test/__mocks__/@polkadot/api.ts
@@ -1,9 +1,0 @@
-import {
-    WsProvider,
-    ApiPromise,
-} from '../PolkadotApiMock';
-
-export {
-    WsProvider,
-    ApiPromise,
-}
