--- packages/node-api/test/__mocks__/@polkadot/extension-dapp.ts
@@ -1,12 +1,0 @@
-import {
-    web3Enable,
-    isWeb3Injected,
-    web3AccountsSubscribe,
-    web3FromAddress,
-} from '../PolkadotExtensionDappMock';
-export {
-    web3Enable,
-    isWeb3Injected,
-    web3AccountsSubscribe,
-    web3FromAddress,
-};
