--- packages/node-api/test/__mocks__/Accounts.ts
@@ -1,3 +1,0 @@
-export function isValidAccountId(): boolean {
-    return true;
-}
