--- packages/node-api/test/__mocks__/Balances.ts
@@ -1,9 +1,0 @@
-import {
-    getBalances,
-    buildTransferCall
-} from './BalancesMock'
-
-export {
-    getBalances,
-    buildTransferCall
-}
