--- packages/node-api/test/__mocks__/BalancesMock.ts
@@ -1,10 +1,0 @@
-
-export let getBalances = jest.fn();
-
-export function setGetBalances(func: any) {
-    getBalances = func
-}
-
-export function buildTransferCall() {
-    return {};
-}
