--- packages/node-api/test/__mocks__/Codec.ts
@@ -1,7 +1,0 @@
-export function toHex(source: string) {
-    return source;
-}
-
-export function fromHex(hex: string) {
-    return hex;
-}
