--- packages/node-api/test/__mocks__/Keys.ts
@@ -1,9 +1,0 @@
-import {
-    isExtensionAvailable,
-    recommendedExtension,
-} from './KeysMock';
-
-export {
-    isExtensionAvailable,
-    recommendedExtension,
-};
