--- packages/node-api/test/__mocks__/KeysMock.ts
@@ -1,19 +1,0 @@
-let extensionAvailable = false;
-
-export function setExtensionAvailable(value: boolean) {
-    extensionAvailable = value;
-}
-
-export function isExtensionAvailable() {
-    return extensionAvailable;
-}
-
-let recommendedExtensionValue: any = null;
-
-export function setRecommendedExtension(value: any) {
-    recommendedExtensionValue = value;
-}
-
-export function recommendedExtension(): any {
-    return recommendedExtensionValue;
-}
