--- packages/node-api/test/__mocks__/LogionLoc.ts
@@ -1,25 +1,0 @@
-import {
-    createLogionIdentityLoc,
-    createLogionTransactionLoc,
-    createPolkadotIdentityLoc,
-    createPolkadotTransactionLoc,
-    getLegalOfficerCase,
-    addMetadata,
-    addFile,
-    addLink,
-    closeLoc,
-    voidLoc,
-} from './LogionLocMock';
-
-export {
-    createLogionIdentityLoc,
-    createLogionTransactionLoc,
-    createPolkadotIdentityLoc,
-    createPolkadotTransactionLoc,
-    getLegalOfficerCase,
-    addMetadata,
-    addFile,
-    addLink,
-    closeLoc,
-    voidLoc,
-}
