--- packages/node-api/test/__mocks__/Recovery.ts
@@ -1,15 +1,0 @@
-import {
-    createRecovery,
-    vouchRecovery,
-    getRecoveryConfig,
-    getActiveRecovery,
-    initiateRecovery,
-} from './RecoveryMock';
-
-export {
-    createRecovery,
-    vouchRecovery,
-    getRecoveryConfig,
-    getActiveRecovery,
-    initiateRecovery,
-}
